# app-mediakit-knowledge/openapi.yaml
openapi: "3.1.0"
info:
  title: app-mediakit-knowledge
  version: "0.2.0"
  description: |
    Wikipedia-pattern HTTP knowledge wiki engine for PointSav Documentation.
    Serves Markdown content at documentation.pointsav.com. Single Rust binary;
    no runtime dependencies beyond the compiled binary.

    Content is read directly from a Git working tree; revision history,
    blame, and diff are served read-only. There is no write path: the engine
    exposes no /auth, /edit, /admin, or /create routes. Editing is performed
    out-of-band through Git, and the running engine renders the committed
    Markdown. This regeneration follows the removal of the optional auth and
    in-browser edit surfaces (Q1: git-only).

    The MCP endpoint (POST /mcp) is mounted only when the binary is started
    with --enable-mcp. When the flag is absent the route is not registered and
    requests return 404.

  contact:
    name: PointSav Digital Systems
    url: https://documentation.pointsav.com
  license:
    name: Apache-2.0
    url: https://opensource.org/licenses/Apache-2.0
servers:
  - url: https://documentation.pointsav.com
    description: Production
...
paths:
...
  /api/history/{slug}:
    get:
      tags: [api, git]
      summary: Article revision history (JSON)
      description: |
        Returns the Git commit history for the given article slug. Each entry
        includes the commit identifier, author, date, and message.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          example: architecture/compounding-substrate
      responses:
        "200":
          description: Revision history
          content:
            application/json:
              schema:
                type: object
                properties:
                  slug:
                    type: string
                  commits:
                    type: array
                    items:
                      type: object
                      properties:
                        sha:
                          type: string
                        author:
                          type: string
                        date:
                          type: string
                          format: date-time
                        message:
                          type: string
        "404":
          description: Article not found
  /api/links/{slug}:
    get:
      tags: [api, git]
      summary: Wikilink graph for an article (JSON)
      description: |
        Returns the forward and backward wikilink edges for the given article
        slug, backed by the in-process link graph.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          example: architecture/compounding-substrate
        - name: direction
          in: query
          required: false
          schema:
            type: string
            enum: [forward, backward, both]
            default: both
          description: "Link direction: forward, backward, or both"
      responses:
        "200":
          description: Wikilink graph edges
          content:
            application/json:
              schema:
                type: object
                properties:
                  slug:
                    type: string
                  forward:
                    type: array
                    items:
                      type: string
                  backward:
                    type: array
                    items:
                      type: string
        "404":
          description: Article not found
...
  /git/{slug}:
    get:
      tags: [git]
      summary: Raw Markdown source (git-clone-style ingestion)
      description: Returns the raw Markdown source for a slug. An optional trailing `.md` suffix is stripped.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          example: architecture/compounding-substrate
      responses:
        "200":
          description: Raw Markdown
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: Slug not found
  /history/{slug}:
    get:
      tags: [git]
      summary: Article revision history (HTML)
      description: Renders the Git commit history for the given article slug as an HTML page.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          example: architecture/compounding-substrate
      responses:
        "200":
          description: History page HTML
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Slug not found
...